+++ prestafford1/life.go
@@ -1,4 +1,4 @@
-package abrash1d
+package prestafford1
 
 import (
 	"math"
@@ -21,6 +21,9 @@ const (
 
 	// We can thus use everything less than that (15) for the count of neighbors.
 	countbit = 0xf
+
+	// Lets store whether the cell's next state in the fifth bit (32). Skipping the syntactic sugar this time.
+	nextbit = 1 << 5
 )
 
 // state gets the state of the cell by AND-ing the fourth bit. The result will be 0 (so, dead) for anything where 00001000 is not true.
@@ -33,9 +36,19 @@ func (c cell) neighbors() cell {
 	return c & countbit
 }
 
+// nextState returns the next state of the cell.
+func (c cell) next() bool {
+	return (c & nextbit) != 0
+}
+
 // vivify makes the cell alive by OR-ing the fourth bit (so, noop if it's already alive).
 func (c cell) vivify() cell {
 	return c | statebit
+}
+
+// vivifyNext sets the cell's next state to alive.
+func (c cell) vivifyNext() cell {
+	return c | nextbit
 }
 
 // kill makes the cell dead by AND-ing the bit with NOT the fourth bit (so, AND-ing with false).
@@ -43,10 +56,16 @@ func (c cell) kill() cell {
 	return c &^ statebit
 }
 
+// killNext sets the cell's next state to dead.
+func (c cell) killNext() cell {
+	return c &^ nextbit
+}
+
 type model struct {
-	width  int
-	height int
-	field  []cell
+	width   int
+	height  int
+	field   []cell
+	changes []int
 }
 
 // wrapPos wraps a cell position that would otherwise be outside of a rectangular grid.
@@ -56,116 +75,147 @@ func (m *model) wrapPos(pos int) int {
 
 // addToNeighbors to all neighboring cells.
 func (m *model) addToNeighbors(pos int) {
+	m.changes = append(m.changes, pos)
+
 	// West
 	newPos := m.wrapPos(pos - 1)
 	m.field[newPos] += 0x1
+	m.changes = append(m.changes, newPos)
 
 	// Northwest
 	newPos = m.wrapPos(pos - m.width - 1)
 	m.field[newPos] += 0x1
+	m.changes = append(m.changes, newPos)
 
 	// North
 	newPos = m.wrapPos(pos - m.width)
 	m.field[newPos] += 0x1
+	m.changes = append(m.changes, newPos)
 
 	// Northeast
 	newPos = m.wrapPos(pos - m.width + 1)
 	m.field[newPos] += 0x1
+	m.changes = append(m.changes, newPos)
 
 	// East
 	newPos = m.wrapPos(pos + 1)
 	m.field[newPos] += 0x1
+	m.changes = append(m.changes, newPos)
 
 	// Southeast
 	newPos = m.wrapPos(pos + m.width + 1)
 	m.field[newPos] += 0x1
+	m.changes = append(m.changes, newPos)
 
 	// South
 	newPos = m.wrapPos(pos + m.width)
 	m.field[newPos] += 0x1
+	m.changes = append(m.changes, newPos)
 
 	// Southwest
 	newPos = m.wrapPos(pos + m.width - 1)
 	m.field[newPos] += 0x1
+	m.changes = append(m.changes, newPos)
 }
 
 // subtractFromNeighbors subtracts from all neighboring cells.
 func (m *model) subtractFromNeighbors(pos int) {
+	m.changes = append(m.changes, pos)
+
 	// West
 	newPos := m.wrapPos(pos - 1)
 	m.field[newPos] -= 0x1
+	m.changes = append(m.changes, newPos)
 
 	// Northwest
 	newPos = m.wrapPos(pos - m.width - 1)
 	m.field[newPos] -= 0x1
+	m.changes = append(m.changes, newPos)
 
 	// North
 	newPos = m.wrapPos(pos - m.width)
 	m.field[newPos] -= 0x1
+	m.changes = append(m.changes, newPos)
 
 	// Northeast
 	newPos = m.wrapPos(pos - m.width + 1)
 	m.field[newPos] -= 0x1
+	m.changes = append(m.changes, newPos)
 
 	// East
 	newPos = m.wrapPos(pos + 1)
 	m.field[newPos] -= 0x1
+	m.changes = append(m.changes, newPos)
 
 	// Southeast
 	newPos = m.wrapPos(pos + m.width + 1)
 	m.field[newPos] -= 0x1
+	m.changes = append(m.changes, newPos)
 
 	// South
 	newPos = m.wrapPos(pos + m.width)
 	m.field[newPos] -= 0x1
+	m.changes = append(m.changes, newPos)
 
 	// Southwest
 	newPos = m.wrapPos(pos + m.width - 1)
 	m.field[newPos] -= 0x1
+	m.changes = append(m.changes, newPos)
 }
 
-// calculateNeighbors calculates alive neighbors for every cell in the model.
+// calculateNeighbors calculates alive neighbors for every cell in the model. It marks all added cells as changed, whether or not they will have, which makes the first generation a little expensive, but that's okay.
 func (m *model) calculateAllNeighbors() {
 	for i, c := range m.field {
 		if c.state() {
 			m.addToNeighbors(i)
+			m.changes = append(m.changes, i)
 		}
 	}
 }
 
-// makeAlive sets the cell state to alive and increments the neighbor count. It assumes the cell is dead.
+// makeAlive sets the cell state to alive and increments the neighbor count. It is idemptotent, assumes the cell is dead, and if it became alive on the current generation, appends the position to the list of changes.
 func (m *model) makeAlive(pos int) {
+	if m.field[pos].state() {
+		return
+	}
 	m.addToNeighbors(pos)
 	m.field[pos] = m.field[pos].vivify()
 }
 
-// makeDead sets the cell state to dead and decrements the neighbor count. It assumes the cell is alive.
+// makeDead sets the cell state to dead and decrements the neighbor count. It is idemptotent, assumes the cell is alive, and if it became dead on the current generation, appends the position to the list of changes.
 func (m *model) makeDead(pos int) {
+	if !m.field[pos].state() {
+		return
+	}
 	m.subtractFromNeighbors(pos)
 	m.field[pos] = m.field[pos].kill()
 }
 
 // nextGeneration evolves the field of automata one generation based on the rules of Conway's Game of Life.
 func (m *model) Next() {
-	// Deep copy the field
-	next := make([]cell, m.width*m.height)
-	for i, _ := range m.field {
-		next[i] = m.field[i]
-	}
+	// Deep copy the field and changelist
+	currentChanges := []int{}
 
 	// Loop over the field.
-	for i, c := range next {
+	for _, change := range m.changes {
+		c := m.field[change]
 
-		// If a cell has zero alive neighbors and is already dead, it's just going to stay dead.
-		if c == 0 {
-			continue
-		}
 		if c.state() {
 			if c.neighbors() != 0x2 && c.neighbors() != 0x3 {
-				m.makeDead(i)
+				currentChanges = append(currentChanges, change)
+				m.field[change] = c.killNext()
 			}
 		} else if c.neighbors() == 0x3 {
-			m.makeAlive(i)
+			currentChanges = append(currentChanges, change)
+			m.field[change] = c.vivifyNext()
+		}
+	}
+	m.changes = []int{}
+	for _, change := range currentChanges {
+		if m.field[change].next() {
+			m.makeAlive(change)
+		} else {
+			m.makeDead(change)
 		}
 	}
 }
@@ -228,9 +278,10 @@ func (m *model) String() string {
 
 func New(width, height int) *model {
 	m := &model{
-		width:  width,
-		height: height,
-		field:  make([]cell, width*height),
+		width:   width,
+		height:  height,
+		field:   make([]cell, width*height),
+		changes: []int{},
 	}
 	return m
 }
